refactor: изменил рендер данных в RecipeCard
