+++ src/App/pages/RecipesPage/RecipesPage.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { RecipesProvider } from 'context/RecipesContext';
+import Hero from './components/Hero';
 import RecipeFilterList from './components/RecipeFilterList';
-import RecipeHero from './components/RecipeHero';
 import RecipeList from './components/RecipeList';
 import RecipePagination from './components/RecipePagination';
 import RecipeSearch from './components/RecipeSearch';
@@ -9,7 +9,7 @@ import RecipeSearch from './components/RecipeSearch';
 const RecipesPage: FC = () => {
   return (
     <RecipesProvider>
-      <RecipeHero />
+      <Hero />
       <RecipeSearch />
       <RecipeFilterList />
       <RecipeList />
