+++ src/App/pages/RecipesPage/components/Hero/Hero.module.scss
@@ -1,4 +1,4 @@
-.item {
+.section {
   width: 100%;
   max-width: 1480px;
   margin: 0 auto;
