+++ src/App/pages/RecipesPage/components/Hero/Hero.tsx
@@ -1,0 +1,8 @@
+import { FC } from 'react';
+import style from './Hero.module.scss';
+
+const Hero: FC = () => {
+  return <div className={style.section} />;
+};
+
+export default Hero;
