--- src/App/pages/RecipesPage/components/RecipeHero/RecipeHero.tsx
@@ -1,8 +1,0 @@
-import { FC } from 'react';
-import style from './RecipeHero.module.scss';
-
-const Hero: FC = () => {
-  return <div className={style.item} />
-};
-
-export default Hero;
