--- src/App/pages/RecipesPage/components/RecipeHero/index.ts
@@ -1,1 +1,0 @@
-export { default } from './RecipeHero';
