+++ src/App/pages/RecipesPage/components/RecipeList/RecipeList.tsx
@@ -1,8 +1,4 @@
 import { FC } from 'react';
-import { useNavigate, generatePath } from 'react-router-dom';
-import BaseButton from 'components/BaseButton';
-import Text from 'components/Text';
-import { ROUTS } from 'config/routs';
 import { useRecipesContext } from 'context/RecipesContext';
 import useFetchRecipeList from '../../hooks/useFetchRecipeList';
 import RecipeCard from '../RecipeCard';
@@ -11,7 +7,6 @@ import style from './RecipeList.module.scss';
 
 const RecipeList: FC = () => {
   const { recipeListState, recipeList, cursorList } = useRecipesContext();
-  const navigate = useNavigate();
   useFetchRecipeList();
 
   return (
@@ -20,30 +15,7 @@ const RecipeList: FC = () => {
         <div className={style.list}>
           {recipeListState === 'loaded' &&
             recipeList.map((recipe) => {
-              const { id, image, readyInMinutes, title, nutrition } = recipe;
-              const cookingTime = `${readyInMinutes} minutes`;
-              const composition = nutrition.ingredients.map(({ name }) => name).join(' + ');
-              const nutritional = `${nutrition.nutrients[0].amount.toFixed()} ${nutrition.nutrients[0].unit}`;
-              const handleClickCard = () => {
-                navigate(generatePath(ROUTS.RECIPE, { id }));
-              };
-
-              return (
-                <RecipeCard
-                  key={id}
-                  image={image}
-                  cookingTime={cookingTime}
-                  title={title}
-                  composition={composition}
-                  nutritional={nutritional}
-                  handleClickCard={handleClickCard}
-                  button={
-                    <BaseButton>
-                      <Text>Save</Text>
-                    </BaseButton>
-                  }
-                />
-              );
+              return <RecipeCard recipe={recipe} key={recipe.id} />;
             })}
           {recipeListState === 'loading' && (
             <div className={style.list}>
